Rename cli -> headless
