+++ pynicotine/headless/__init__.py
@@ -17,8 +17,8 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 
-def run_cli(core, ci_mode):
-    """ Run application in headless (CLI) mode """
+def run_headless(core, ci_mode):
+    """ Run application in headless (no GUI) mode """
 
-    from pynicotine.cli.application import Application
+    from pynicotine.headless.application import Application
     return Application(core, ci_mode).run()
