+++ pynicotine/headless/application.py
